delay load ptblocks
