+++ webapp/public/index.html
@@ -48,7 +48,6 @@
     <script type="text/javascript" src="/blb/smoothie/smoothie_compressed.js"></script>
     <script type="text/javascript" src="/blb/semantic.js"></script>
     <script type="text/javascript" src="/blb/pxtlib.js"></script>
-    <script type="text/javascript" src="/blb/pxtblocks.js"></script>
     <script type="text/javascript" src="/blb/pxtwinrt.js"></script>
     <script type="text/javascript" src="/blb/pxteditor.js"></script>
     <script type="text/javascript" src="/blb/pxtsim.js"></script>
@@ -90,7 +89,8 @@
             "lzma/lzma_worker-min.js": "/blb/lzma/lzma_worker-min.js",
             "blockly/blockly_compressed.js": "/blb/blockly/blockly_compressed.js",
             "blockly/blocks_compressed.js": "/blb/blockly/blocks_compressed.js",
-            "blockly/msg/js/en.js": "/blb/blockly/msg/js/en.js"
+            "blockly/msg/js/en.js": "/blb/blockly/msg/js/en.js",
+            "pxtblocks.js": "/blb/pxtblocks.js"
         }
     </script>
 </body>
