+++ webapp/src/monaco.tsx
@@ -108,7 +108,7 @@ export class Editor extends srceditor.Editor {
             // 3) check that decompiled js == current js % white space
             let blocksInfo: pxtc.BlocksInfo;
             return this.parent.saveFileAsync()
-                .then(() => pxt.blocks.loadBlocklyAsync())
+                .then(() => pxt.BrowserUtils.loadBlocklyAsync())
                 .then(() => compiler.getBlocksAsync())
                 .then((bi: pxtc.BlocksInfo) => {
                     blocksInfo = bi;
